@@ -19,7 +19,7 @@
 
 package org.akvo.caddisfly.sensor.colorimetry.strip.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.HashMap;
