@@ -23,8 +23,9 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.pm.PackageManager;
 import android.hardware.Camera;
-import android.support.annotation.StringRes;
 import android.util.Log;
+
+import androidx.annotation.StringRes;
 
 import org.akvo.caddisfly.R;
 import org.akvo.caddisfly.util.AlertUtil;
