@@ -21,9 +21,10 @@ package org.akvo.caddisfly.util;
 
 import android.graphics.Bitmap;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
 import android.util.Log;
 import android.util.SparseIntArray;
+
+import androidx.annotation.NonNull;
 
 import org.akvo.caddisfly.model.ColorInfo;
 import org.akvo.caddisfly.model.LabColor;
